requirements.txt
ccxt
pandas
numpy
gymnasium
stable-baselines3
matplotlib
optuna
pyyaml
